refactor common CUDA/HIP code into common header
